automated_egg_incubator/requirements.txt
requests
Adafruit_DHT
RPi.GPIO as GPIO
ConfigParser
